# packages/contracts/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoCoLo API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/feature-contract": {
      "get": {
        "operationId": "getFeatureContract",
        "summary": "選択中チームの有効機能契約を取得",
        "responses": {
          "200": {
            "description": "有効機能契約",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureContractResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/feature-contract/{featureKey}": {
      "patch": {
        "operationId": "updateFreeFeatureFlag",
        "summary": "チームの無償feature flagを変更",
        "parameters": [
          {
            "name": "featureKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9._-]{1,63}$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureFlagUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "feature flag変更後の有効機能契約",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureContractResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "FeatureFlagUpdate": {
        "type": "object",
        "required": [
          "enabled",
          "reason"
        ],
        "additionalProperties": false,
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          }
        }
      },
      "FeatureContractResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "planKey",
              "planStatus",
              "features"
            ],
            "additionalProperties": false,
            "properties": {
              "planKey": {
                "type": [
                  "string",
                  "null"
                ],
                "maxLength": 100
              },
              "planStatus": {
                "type": [
                  "string",
                  "null"
                ],
                "enum": [
                  "active",
                  "trialing",
                  "past_due",
                  "canceled",
                  "expired",
                  null
                ]
              },
              "features": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": [
                    "key",
                    "billingType",
                    "displayName",
                    "enabled",
                    "reason"
                  ],
                  "additionalProperties": false,
                  "properties": {
                    "key": {
                      "type": "string",
                      "pattern": "^[a-z][a-z0-9_.-]{0,63}$"
                    },
                    "billingType": {
                      "type": "string",
                      "enum": [
                        "free",
                        "paid"
                      ]
                    },
                    "displayName": {
                      "type": "string",
                      "minLength": 1,
                      "maxLength": 200
                    },
                    "enabled": {
                      "type": "boolean"
                    },
                    "reason": {
                      "type": "string",
                      "enum": [
                        "default",
                        "flag",
                        "plan",
                        "unavailable"
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              "details": {},
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "入力値が不正です。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "認証が必要です。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "この操作を実行する権限がありません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "指定されたリソースが見つかりません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "リクエスト数の上限を超えました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "内部エラーが発生しました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "外部サービスまたは依存サービスが利用できません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
...
}